+++ app-android/src/main/java/com/kplc/outage/android/di/AndroidModule.kt
@@ -1,0 +1,6 @@
+package com.kplc.outage.android.di
+
+import org.koin.dsl.module
+
+val androidModule = module {
+}
