--- app-android/src/main/java/com/vickikbt/kmptemplate/android/di/AndroidModule.kt
@@ -1,6 +1,0 @@
-package com.vickikbt.kmptemplate.android.di
-
-import org.koin.dsl.module
-
-val androidModule = module {
-}
