+++ shared/src/androidMain/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.vickikbt.kmptemplate" />
+<manifest package="com.kplc.outage" />
