+++ shared/src/commonMain/kotlin/com/kplc/outage/di/Koin.kt
@@ -1,0 +1,12 @@
+package com.kplc.outage.di
+
+import org.koin.core.context.startKoin
+import org.koin.dsl.KoinAppDeclaration
+
+fun initKoin(isDebug: Boolean = false, appDeclaration: KoinAppDeclaration = {}) =
+    startKoin {
+        appDeclaration()
+        modules(commonModule(isDebug = isDebug))
+    }
+
+fun initKoin() = initKoin {}
