+++ shared/src/commonMain/kotlin/com/kplc/outage/domain/utils/Platform.kt
@@ -1,0 +1,5 @@
+package com.kplc.outage.domain.utils
+
+expect class Platform() {
+    val platform: String
+}
