+++ shared/src/commonMain/kotlin/com/kplc/outage/presentation/utils/Greeting.kt
@@ -1,0 +1,9 @@
+package com.kplc.outage.presentation.utils
+
+import com.kplc.outage.domain.utils.Platform
+
+class Greeting {
+    fun greet(): String {
+        return "Hello, ${Platform().platform}!"
+    }
+}
