--- shared/src/commonMain/kotlin/com/vickikbt/kmptemplate/domain/utils/Platform.kt
@@ -1,5 +1,0 @@
-package com.vickikbt.kmptemplate.domain.utils
-
-expect class Platform() {
-    val platform: String
-}
