+++ shared/src/jvmMain/kotlin/com/kplc/outage/domain/utils/Platform.kt
@@ -1,0 +1,5 @@
+package com.kplc.outage.domain.utils
+
+actual class Platform actual constructor() {
+    actual val platform: String = "Desktop"
+}
