--- shared/src/jvmMain/kotlin/com/vickikbt/kmptemplate/domain/utils/Platform.kt
@@ -1,5 +1,0 @@
-package com.vickikbt.kmptemplate.domain.utils
-
-actual class Platform actual constructor() {
-    actual val platform: String = "Desktop"
-}
